backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.9.2
pillow==10.4.0
qrcode[pil]==7.4.2
shortuuid==1.0.13
